Test for hotel search, car search, car details and other minor changes

+++ src/main/java/com/models/allmyles/hotels/details/CityCenter.java
@@ -4,15 +4,15 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class CityCenter {
-    double distance;
+    float distance;
     String unit;
 
 
-    public double getDistance() {
+    public float getDistance() {
         return distance;
     }
 
-    public void setDistance(double distance) {
+    public void setDistance(float distance) {
         this.distance = distance;
     }
